+++ packages/ui-kit/src/components/TableCanvas/TableCanvas.tsx
@@ -1005,6 +1005,15 @@ export function TableCanvas<
                               maxColumnAutoWidth:
                                 tableConfig.maxColumnAutoWidth,
                               resizableColumn: tableConfig.resizableColumn,
+                              spanGroupHeader:
+                                tableConfig.columnsGrouping?.squashEmptyCells ??
+                                true,
+                              spanShallowGroups:
+                                tableConfig.columnsGrouping?.squashEmptyCells ??
+                                true,
+                              spanAlign:
+                                tableConfig.columnsGrouping
+                                  ?.squashedHeaderAlign,
                               editorOverlayPortal:
                                 tableConfig.editorOverlayPortal,
                               checkboxSelectedRowIndexes,